+++ packages/agent/src/index.ts
@@ -1,4 +1,4 @@
-import { CollectionCustomizer, TSchema } from '@forestadmin/datasource-customizer';
+import { TSchema } from '@forestadmin/datasource-customizer';
 
 import Agent from './agent';
 import { AgentOptions } from './types';
@@ -8,8 +8,8 @@ export function createAgent<S extends TSchema = TSchema>(options: AgentOptions):
 }
 
 export { Agent };
-export { CollectionCustomizer };
 export { AgentOptions } from './types';
+export * from '@forestadmin/datasource-customizer';
 
 // export is necessary for the agent-generator package
 export { default as SchemaGenerator } from './utils/forest-schema/generator';