+++ packages/agent/src/services/authorization/types.ts
@@ -18,6 +18,16 @@ export interface SmartActionRequestBody {
   };
 }
 
+export interface SmartActionHookRequestBody extends SmartActionRequestBody {
+  data: SmartActionRequestBody['data'] & {
+    attributes: SmartActionRequestBody['data']['attributes'] & {
+      changed_field: string;
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      fields: any[];
+    };
+  };
+}
+
 export interface SmartActionApprovalRequestBody extends SmartActionRequestBody {
   data: SmartActionRequestBody['data'] & {
     attributes: SmartActionRequestBody['data']['attributes'] & {