+++ packages/datasource-customizer/src/collection-customizer.ts
@@ -97,7 +97,7 @@ export default class CollectionCustomizer<
   }
 
   /**
-   * Remove field by setting its visibility to false.
+   * Remove fields from the exported schema (they will still be usable within the agent).
    * @param names the fields to remove
    * @see {@link https://docs.forestadmin.com/developer-guide-agents-nodejs/agent-customization/fields/import-rename-delete#renaming-and-removing-fields Documentation Link}
    * @example
@@ -138,11 +138,12 @@ export default class CollectionCustomizer<
    * @param definition definition of the chart
    * @see {@link https://docs.forestadmin.com/developer-guide-agents-nodejs/agent-customization/charts Documentation Link}
    * @example
-   * .addChart('numCustomers', {
-   *   type: 'Value',
-   *   render: (context, resultBuilder) => {
-   *     return resultBuilder.value(123);
-   *   }
+   * .addChart('numCustomers', (context, resultBuilder) => {
+   *   return resultBuilder.distribution({
+   *     tomatoes: 10,
+   *     potatoes: 20,
+   *     carrots: 30,
+   *   });
    * })
    */
   addChart(name: string, definition: CollectionChartDefinition<S, N>): this {