+++ packages/datasource-toolkit/test/decorators/collection-decorator.test.ts
@@ -206,10 +206,10 @@ describe('CollectionDecorator', () => {
       );
 
       const caller = factories.caller.build();
-      const result = await decoratedCollection.getForm(caller, 'an action name', {}, null);
+      const result = await decoratedCollection.getForm(caller, 'an action name', {}, null, null);
 
       expect(result).toStrictEqual(fields);
-      expect(childGetForm).toHaveBeenCalledWith(caller, 'an action name', {}, null);
+      expect(childGetForm).toHaveBeenCalledWith(caller, 'an action name', {}, null, null);
     });
   });