+++ packages/plugin-aws-s3/CHANGELOG.md
@@ -1,3 +1,45 @@
+## @forestadmin/plugin-aws-s3 [1.3.6](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.5...@forestadmin/plugin-aws-s3@1.3.6) (2023-07-26)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.13.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.7.0
+
+## @forestadmin/plugin-aws-s3 [1.3.5](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.4...@forestadmin/plugin-aws-s3@1.3.5) (2023-07-26)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.12.1
+
+## @forestadmin/plugin-aws-s3 [1.3.4](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.3...@forestadmin/plugin-aws-s3@1.3.4) (2023-07-26)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.12.0
+
+## @forestadmin/plugin-aws-s3 [1.3.3](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.2...@forestadmin/plugin-aws-s3@1.3.3) (2023-07-25)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.11.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.6.0
+
 ## @forestadmin/plugin-aws-s3 [1.3.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.1...@forestadmin/plugin-aws-s3@1.3.2) (2023-07-20)